requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
torch>=2.2.0
numpy>=1.26.0
supabase>=2.4.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
pytest>=8.0.0
httpx>=0.27.0
scikit-learn>=1.4.0
